--- application-bemyplan/src/main/java/com/deploy/bemyplan/service/auth/AuthService.java
@@ -1,10 +1,0 @@
-package com.deploy.bemyplan.service.auth;
-
-import com.deploy.bemyplan.service.auth.dto.request.LoginDto;
-import com.deploy.bemyplan.service.auth.dto.request.SignUpDto;
-
-public interface AuthService {
-
-    Long signUp(SignUpDto request);
-    Long login(LoginDto request);
-}
