+++ applications/app-bemyplan/src/main/java/com/deploy/bemyplan/auth/controller/dto/request/SignOutUserRequest.java
@@ -1,6 +1,10 @@
-package com.deploy.bemyplan.controller.auth.dto.request;
+package com.deploy.bemyplan.auth.controller.dto.request;
 
-import lombok.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 import javax.validation.constraints.NotBlank;
